requirements.txt
Langchain
Langchain-community
ollama
pandas
openpyxl
PyGithub
llama-index